# vulkan: Pad N dimension of B matrix for coopmat2 perf, to avoid bounds checking